requirements.txt
opencv-python
blurhash-python
tqdm
